+++ rugui/src/lib.rs
@@ -6,5 +6,4 @@
 pub mod coordinates;
 pub mod framebuffer;
 pub mod geometry;
-mod geometry_tests;
 pub mod widgets;
