+++ rugui/src/widgets/mod.rs
@@ -1,0 +1,3 @@
+pub mod progress_bar;
+pub mod scroller;
+pub mod table;
